Write strings to buffers
Rather than printing a character at a time, `print_int` and `print_uint`
will now write the representation to a stack-allocated buffer and use
the `print_string` function to display the result.

This should be much more performant and is also inline with the
expectations of the textbook exercise this functionality is built for.